# docs/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/ideas:
    get:
      operationId: IdeasController_list
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a1
        - Ideas
    post:
      operationId: IdeasController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdeaDto"
      responses:
        "201":
          description: ""
      tags: *a1
...
  /v1/plans:
    get:
      operationId: PlansController_list
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a2
        - Plans
    post:
      operationId: PlansController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanDto"
      responses:
        "201":
          description: ""
      tags: *a2
  /v1/plans/{planId}:
    get:
      operationId: PlansController_get
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a2
    patch:
      operationId: PlansController_update
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanDto"
      responses:
        "200":
          description: ""
      tags: *a2
    delete:
      operationId: PlansController_remove
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags: *a2
...
  /v1/plans/{planId}/sections/{sectionId}:
    patch:
      operationId: PlansController_updateSection
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
        - name: sectionId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectionDto"
      responses:
        "200":
          description: ""
      tags: *a2
    delete:
      operationId: PlansController_removeSection
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
        - name: sectionId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags: *a2
...
info:
  title: IREAL Service API
  description: Versioned API for the decoupled IREAL backend service.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  schemas:
    CreateIdeaDto:
      type: object
      properties: {}
...
    CreatePlanDto:
      type: object
      properties: {}
    UpdatePlanDto:
      type: object
      properties: {}
...
    UpdateSectionDto:
      type: object
      properties: {}
...